Simplifying the specialized setIfInInput
