+++ src/Estimators/OneBodyDensityMatricesInput.cpp
@@ -2,7 +2,7 @@
 // This file is distributed under the University of Illinois/NCSA Open Source License.
 // See LICENSE file in top directory for details.
 //
-// Copyright (c) 2021 QMCPACK developers.
+// Copyright (c) 2023 QMCPACK developers.
 //
 // File developed by: Peter Doak, doakpw@ornl.gov, Oak Ridge National Laboratory
 //
@@ -12,7 +12,7 @@
 #include "string_utils.h"
 #include "EstimatorInput.h"
 #include "OneBodyDensityMatricesInput.h"
-#include "SetIfInInput.hpp"
+
 namespace qmcplusplus
 {
