+++ src/Estimators/ReferencePointsInput.cpp
@@ -2,7 +2,7 @@
 // This file is distributed under the University of Illinois/NCSA Open Source License.
 // See LICENSE file in top directory for details.
 //
-// Copyright (c) 2022 QMCPACK developers.
+// Copyright (c) 2023 QMCPACK developers.
 //
 // File developed by: Jaron T. Krogel, krogeljt@ornl.gov, Oak Ridge National Laboratory
 //                    Mark A. Berrill, berrillma@ornl.gov, Oak Ridge National Laboratory
@@ -17,7 +17,6 @@
 
 #include "EstimatorInput.h"
 #include "ModernStringUtils.hpp"
-#include "SetIfInInput.hpp"
 
 namespace qmcplusplus
 {
