+++ src/Estimators/SpaceGridInput.h
@@ -2,7 +2,7 @@
 // This file is distributed under the University of Illinois/NCSA Open Source License.
 // See LICENSE file in top directory for details.
 //
-// Copyright (c) 2022 QMCPACK developers.
+// Copyright (c) 2023 QMCPACK developers.
 //
 // File developed by: Peter Doak, doakpw@ornl.gov, Oak Ridge National Laboratory
 //
@@ -19,7 +19,6 @@
 #include "InputSection.h"
 #include "EstimatorInput.h"
 #include "ParseGridInput.hpp" //for AxisGrid
-#include "SetIfInInput.hpp"
 
 namespace qmcplusplus
 {
